feat: use continuous fitness scoring and add exp fixture

+++ cpp/include/g3pvm/vm_cpu.hpp
@@ -28,12 +28,12 @@ VMResult run_bytecode(const BytecodeProgram& program,
 
 // Returns one fitness score per program.
 // Scoring:
-//   score = exact_match_count - round(mean_abs_error) + runtime_error_count * (-10)
+//   score = exact_match_count - mean_abs_error + runtime_error_count * (-10)
 //           - non_numeric_mismatch_count
 // where mean_abs_error is over numeric predictions/targets across all cases.
 // Non-numeric non-error mismatches are penalized by -1 each.
 // All programs share one case set.
-std::vector<int> run_bytecode_cpu_multi_fitness_shared_cases(
+std::vector<double> run_bytecode_cpu_multi_fitness_shared_cases(
     const std::vector<BytecodeProgram>& programs,
     const std::vector<InputCase>& shared_cases,
     const std::vector<Value>& shared_answer,
